Implementert Supervisor

+++ Network_Supervisor/Supervisor/counter.go
@@ -1,10 +1,10 @@
 package supervisor
 
-func IsNewer(newCounter, oldCounter uint8) bool{
+func IsNewer(newCounter, oldCounter uint8) bool {
 	delta := newCounter - oldCounter
-	return delta > 0 && delta < 128 //Half-wrapper, altså for stor forskjell tolereres ikke 
+	return delta > 0 && delta < 128 //Half-wrapper, altså for stor forskjell tolereres ikke
 }
 
-func Delta(newCounter, oldCounter uint8) uint8 {
-	return newCounter - oldCounter
+func Delta(newCounter, oldCounter uint8) int {
+	return int(newCounter - oldCounter)
 }
